Remove unnecessary code for DomainExtractor
