+++ presto-main/src/main/java/com/facebook/presto/sql/relational/DomainExtractor.java
@@ -21,12 +21,8 @@ import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.function.FunctionHandle;
 import com.facebook.presto.metadata.FunctionMetadata;
 import com.facebook.presto.spi.function.OperatorType;
-import com.facebook.presto.spi.predicate.DiscreteValues;
 import com.facebook.presto.spi.predicate.Domain;
-import com.facebook.presto.spi.predicate.Marker;
 import com.facebook.presto.spi.predicate.Range;
-import com.facebook.presto.spi.predicate.Ranges;
-import com.facebook.presto.spi.predicate.SortedRangeSet;
 import com.facebook.presto.spi.predicate.TupleDomain;
 import com.facebook.presto.spi.predicate.Utils;
 import com.facebook.presto.spi.predicate.ValueSet;
@@ -44,18 +40,15 @@ import com.facebook.presto.sql.planner.RowExpressionInterpreter;
 import com.facebook.presto.sql.planner.optimizations.RowExpressionCanonicalizer;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.PeekingIterator;
 
 import javax.annotation.Nullable;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
 import static com.facebook.presto.metadata.CastType.CAST;
 import static com.facebook.presto.metadata.CastType.SATURATED_FLOOR_CAST;
-import static com.facebook.presto.spi.function.OperatorType.BETWEEN;
 import static com.facebook.presto.spi.function.OperatorType.EQUAL;
 import static com.facebook.presto.spi.function.OperatorType.GREATER_THAN;
 import static com.facebook.presto.spi.function.OperatorType.GREATER_THAN_OR_EQUAL;
@@ -66,213 +59,24 @@ import static com.facebook.presto.spi.function.OperatorType.NOT_EQUAL;
 import static com.facebook.presto.spi.predicate.TupleDomain.columnWiseUnion;
 import static com.facebook.presto.spi.relation.SpecialFormExpression.Form.AND;
 import static com.facebook.presto.spi.relation.SpecialFormExpression.Form.IN;
-import static com.facebook.presto.spi.relation.SpecialFormExpression.Form.IS_NULL;
 import static com.facebook.presto.spi.relation.SpecialFormExpression.Form.OR;
 import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
 import static com.facebook.presto.sql.analyzer.TypeSignatureProvider.fromTypes;
 import static com.facebook.presto.sql.planner.LiteralEncoder.toRowExpression;
 import static com.facebook.presto.sql.relational.DomainExtractor.ExtractionResult.resultOf;
 import static com.facebook.presto.sql.relational.Expressions.call;
-import static com.facebook.presto.sql.relational.LogicalRowExpressions.FALSE;
 import static com.facebook.presto.sql.relational.LogicalRowExpressions.TRUE;
 import static com.facebook.presto.sql.relational.LogicalRowExpressions.and;
 import static com.facebook.presto.sql.relational.LogicalRowExpressions.or;
 import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.collect.ImmutableList.toImmutableList;
-import static com.google.common.collect.Iterables.getOnlyElement;
-import static com.google.common.collect.Iterators.peekingIterator;
 import static java.util.Objects.requireNonNull;
-import static java.util.stream.Collectors.collectingAndThen;
 import static java.util.stream.Collectors.partitioningBy;
-import static java.util.stream.Collectors.toList;
 
 public class DomainExtractor
 {
-    private final FunctionManager functionManager;
-    private final LogicalRowExpressions logicalRowExpressions;
-    private final StandardFunctionResolution functionResolution;
-
-    public DomainExtractor(FunctionManager functionManager)
+    private DomainExtractor()
     {
-        this.functionManager = requireNonNull(functionManager, "functionManager is null");
-        this.logicalRowExpressions = new LogicalRowExpressions(functionManager);
-        this.functionResolution = new StandardFunctionResolution(functionManager);
-    }
-
-    public RowExpression toPredicate(TupleDomain<? extends RowExpression> tupleDomain)
-    {
-        if (tupleDomain.isNone()) {
-            return FALSE;
-        }
-
-        Map<? extends RowExpression, Domain> domains = tupleDomain.getDomains().get();
-        return domains.entrySet().stream()
-                .map(entry -> toPredicate(entry.getValue(), entry.getKey()))
-                .collect(collectingAndThen(toImmutableList(), logicalRowExpressions::combineConjuncts));
-    }
-
-    private RowExpression toPredicate(Domain domain, RowExpression reference)
-    {
-        if (domain.getValues().isNone()) {
-            return domain.isNullAllowed() ? isNull(reference) : FALSE;
-        }
-
-        if (domain.getValues().isAll()) {
-            return domain.isNullAllowed() ? TRUE : not(functionResolution, isNull(reference));
-        }
-
-        List<RowExpression> disjuncts = new ArrayList<>();
-
-        disjuncts.addAll(domain.getValues().getValuesProcessor().transform(
-                ranges -> extractDisjuncts(domain.getType(), ranges, reference),
-                discreteValues -> extractDisjuncts(domain.getType(), discreteValues, reference),
-                allOrNone -> {
-                    throw new IllegalStateException("Case should not be reachable");
-                }));
-
-        // Add nullability disjuncts
-        if (domain.isNullAllowed()) {
-            disjuncts.add(isNull(reference));
-        }
-
-        return logicalRowExpressions.combineDisjunctsWithDefault(disjuncts, TRUE);
-    }
-
-    private RowExpression processRange(Type type, Range range, RowExpression reference)
-    {
-        if (range.isAll()) {
-            return TRUE;
-        }
-
-        if (isBetween(range)) {
-            // specialize the range with BETWEEN expression if possible b/c it is currently more efficient
-            return call(
-                    functionManager.resolveOperator(BETWEEN, fromTypes(reference.getType(), type, type)),
-                    BOOLEAN,
-                    reference,
-                    toRowExpression(range.getLow().getValue(), type),
-                    toRowExpression(range.getHigh().getValue(), type));
-        }
-
-        List<RowExpression> rangeConjuncts = new ArrayList<>();
-        if (!range.getLow().isLowerUnbounded()) {
-            switch (range.getLow().getBound()) {
-                case ABOVE:
-                    rangeConjuncts.add(greaterThan(reference, toRowExpression(range.getLow().getValue(), type)));
-                    break;
-                case EXACTLY:
-                    rangeConjuncts.add(greaterThanOrEqual(reference, toRowExpression(range.getLow().getValue(), type)));
-                    break;
-                case BELOW:
-                    throw new IllegalStateException("Low Marker should never use BELOW bound: " + range);
-                default:
-                    throw new AssertionError("Unhandled bound: " + range.getLow().getBound());
-            }
-        }
-        if (!range.getHigh().isUpperUnbounded()) {
-            switch (range.getHigh().getBound()) {
-                case ABOVE:
-                    throw new IllegalStateException("High Marker should never use ABOVE bound: " + range);
-                case EXACTLY:
-                    rangeConjuncts.add(lessThanOrEqual(reference, toRowExpression(range.getHigh().getValue(), type)));
-                    break;
-                case BELOW:
-                    rangeConjuncts.add(lessThan(reference, toRowExpression(range.getHigh().getValue(), type)));
-                    break;
-                default:
-                    throw new AssertionError("Unhandled bound: " + range.getHigh().getBound());
-            }
-        }
-        // If rangeConjuncts is null, then the range was ALL, which should already have been checked for
-        checkState(!rangeConjuncts.isEmpty());
-        return logicalRowExpressions.combineConjuncts(rangeConjuncts);
-    }
-
-    private RowExpression combineRangeWithExcludedPoints(Type type, RowExpression reference, Range range, List<RowExpression> excludedPoints)
-    {
-        if (excludedPoints.isEmpty()) {
-            return processRange(type, range, reference);
-        }
-
-        RowExpression excludedPointsExpression = not(functionResolution, in(reference, excludedPoints));
-        if (excludedPoints.size() == 1) {
-            excludedPointsExpression = notEqual(reference, getOnlyElement(excludedPoints));
-        }
-
-        return logicalRowExpressions.combineConjuncts(processRange(type, range, reference), excludedPointsExpression);
-    }
-
-    private List<RowExpression> extractDisjuncts(Type type, Ranges ranges, RowExpression reference)
-    {
-        List<RowExpression> disjuncts = new ArrayList<>();
-        List<RowExpression> singleValues = new ArrayList<>();
-        List<Range> orderedRanges = ranges.getOrderedRanges();
-
-        SortedRangeSet sortedRangeSet = SortedRangeSet.copyOf(type, orderedRanges);
-        SortedRangeSet complement = sortedRangeSet.complement();
-
-        List<Range> singleValueExclusionsList = complement.getOrderedRanges().stream().filter(Range::isSingleValue).collect(toList());
-        List<Range> originalUnionSingleValues = SortedRangeSet.copyOf(type, singleValueExclusionsList).union(sortedRangeSet).getOrderedRanges();
-        PeekingIterator<Range> singleValueExclusions = peekingIterator(singleValueExclusionsList.iterator());
-
-        for (Range range : originalUnionSingleValues) {
-            if (range.isSingleValue()) {
-                singleValues.add(toRowExpression(range.getSingleValue(), type));
-                continue;
-            }
-
-            // attempt to optimize ranges that can be coalesced as long as single value points are excluded
-            List<RowExpression> singleValuesInRange = new ArrayList<>();
-            while (singleValueExclusions.hasNext() && range.contains(singleValueExclusions.peek())) {
-                singleValuesInRange.add(toRowExpression(singleValueExclusions.next().getSingleValue(), type));
-            }
-
-            if (!singleValuesInRange.isEmpty()) {
-                disjuncts.add(combineRangeWithExcludedPoints(type, reference, range, singleValuesInRange));
-                continue;
-            }
-
-            disjuncts.add(processRange(type, range, reference));
-        }
-
-        // Add back all of the possible single values either as an equality or an IN predicate
-        if (singleValues.size() == 1) {
-            disjuncts.add(equal(reference, getOnlyElement(singleValues)));
-        }
-        else if (singleValues.size() > 1) {
-            disjuncts.add(in(reference, singleValues));
-        }
-        return disjuncts;
-    }
-
-    private List<RowExpression> extractDisjuncts(Type type, DiscreteValues discreteValues, RowExpression reference)
-    {
-        List<RowExpression> values = discreteValues.getValues().stream()
-                .map(object -> toRowExpression(object, type))
-                .collect(toList());
-
-        // If values is empty, then the equatableValues was either ALL or NONE, both of which should already have been checked for
-        checkState(!values.isEmpty());
-
-        RowExpression predicate;
-        if (values.size() == 1) {
-            predicate = equal(reference, getOnlyElement(values));
-        }
-        else {
-            predicate = in(reference, values);
-        }
-
-        if (!discreteValues.isWhiteList()) {
-            predicate = not(functionResolution, predicate);
-        }
-        return ImmutableList.of(predicate);
-    }
-
-    private static boolean isBetween(Range range)
-    {
-        return !range.getLow().isLowerUnbounded() && range.getLow().getBound() == Marker.Bound.EXACTLY
-                && !range.getHigh().isUpperUnbounded() && range.getHigh().getBound() == Marker.Bound.EXACTLY;
     }
 
     /**
@@ -807,19 +611,6 @@ public class DomainExtractor
         }
     }
 
-    private static List<RowExpression> extractColumns(TupleDomain<RowExpression> tupleDomain)
-    {
-        return tupleDomain.getColumnDomains().orElse(ImmutableList.of())
-                .stream()
-                .map(TupleDomain.ColumnDomain::getColumn)
-                .collect(toImmutableList());
-    }
-
-    private static RowExpression isNull(RowExpression expression)
-    {
-        return new SpecialFormExpression(IS_NULL, BOOLEAN, expression);
-    }
-
     private static RowExpression not(StandardFunctionResolution resolution, RowExpression expression)
     {
         return call("not", resolution.notFunction(), expression.getType(), expression);
@@ -828,41 +619,6 @@ public class DomainExtractor
     private static RowExpression in(RowExpression value, List<RowExpression> inList)
     {
         return new SpecialFormExpression(IN, BOOLEAN, ImmutableList.<RowExpression>builder().add(value).addAll(inList).build());
-    }
-
-    private RowExpression binaryOperator(OperatorType operatorType, RowExpression left, RowExpression right)
-    {
-        return call(functionManager.resolveOperator(operatorType, fromTypes(left.getType(), right.getType())), BOOLEAN, left, right);
-    }
-
-    private RowExpression greaterThan(RowExpression left, RowExpression right)
-    {
-        return binaryOperator(OperatorType.GREATER_THAN, left, right);
-    }
-
-    private RowExpression lessThan(RowExpression left, RowExpression right)
-    {
-        return binaryOperator(OperatorType.LESS_THAN, left, right);
-    }
-
-    private RowExpression greaterThanOrEqual(RowExpression left, RowExpression right)
-    {
-        return binaryOperator(GREATER_THAN_OR_EQUAL, left, right);
-    }
-
-    private RowExpression lessThanOrEqual(RowExpression left, RowExpression right)
-    {
-        return binaryOperator(LESS_THAN_OR_EQUAL, left, right);
-    }
-
-    private RowExpression equal(RowExpression left, RowExpression right)
-    {
-        return binaryOperator(EQUAL, left, right);
-    }
-
-    private RowExpression notEqual(RowExpression left, RowExpression right)
-    {
-        return binaryOperator(NOT_EQUAL, left, right);
     }
 
     private static class NormalizedSimpleComparison
@@ -909,11 +665,6 @@ public class DomainExtractor
                 rest = new SpecialFormExpression(AND, BOOLEAN, results);
             }
             return new ExtractionResult(TupleDomain.all(), rest);
-        }
-
-        public static ExtractionResult resultOf(TupleDomain<RowExpression> tupleDomain)
-        {
-            return new ExtractionResult(tupleDomain, TRUE);
         }
 
         public static ExtractionResult resultOf(RowExpression key, Domain domain)
